Fixing imports:
Fixed missing imports in service.py. It looks like those got unnoticed and only broke the GTK GUI.

interface.py: gi.require_version() raises a ValueError, if the specified module is not available.
Catch this in addition to the ImportError and mark ATSPI as unavailable in this case, too.

Fixed imports in GTK GUI.

Bugfix in iomediator package: CURRENT_INTERFACE is a global variable and must not be imported like 'from x import name'