Qt: CentralWidget: Added _get_current_treewidget_item()
This method returns the tree widget item widget corresponding to the
currently edited item. Normally, this is just the selected item, but
if the user deselects it (which can’t be prevented?), the item has
to be retrieved differently. This function simply scans the whole tree
to find the relevant item.

Fixes #285