Fix clear button not removing hotkey (#550)
Previously, clicking "clear" does not change how the hotkeys are
persisted, but it does remove TriggerMode.HOTKEY from the item, which is
persisted as an empty "modes" array in the json. When autokey is
reloaded, the empty "modes" array means the hotkeys in that item are
ignored. With the aforementioned lines adding that triggermode back, the
hotkeys that are persisted are seen and taken up as hotkeys.

Priviously, saving in the gui didn't touch the hotkey if the hotkey box
wasn't enabled.
Now: saving in the gui calls item.unset_hotkey if hotkey box is
disabled.