deps: Add pyasynccore
It looks like asyncore was removed from stdlib in 3.12. So, in it's
place there is a drop-in replacement called pyasyncore available at
[simonrob/pyasyncore](https://github.com/simonrob/pyasyncore).

That project advises new projects to make use of [asyncio](https://docs.python.org/3/library/asyncio.html)
instead, but I assume that would be too big of a lift for the
existing codebase here.